Explicit type cast to avoid compiler warnings

+++ LibSWBF2/MATL.cpp
@@ -16,7 +16,7 @@ namespace LibSWBF2::Chunks::Mesh
 	void MATL::RefreshSize()
 	{
 		ChunkSize matChunksSize = 0;
-		for (int i = 0; i < m_Materials.size(); ++i)
+		for (size_t i = 0; i < m_Materials.size(); ++i)
 		{
 			m_Materials[i].RefreshSize();
 			matChunksSize += m_Materials[i].GetSize();
